Coalesce into single file
